Pulled in 'tip' of lighblue bean library...
Kinda works now...

+++ app/src/main/java/com/ndipatri/roboButton/dagger/RBModule.java
@@ -94,7 +94,7 @@ public class RBModule {
 
         // This is the number of bytes into the BLE 'scanRecord' where the 16-byte
         // UUID begins.
-        Integer[] regionUUIDOffsetArray = new Integer[] {new Integer(2),
+        Integer[] regionUUIDOffsetArray = new Integer[] {new Integer(3),
                                                          new Integer(0),
                                                         };
